Add Real-Time GigEVision device and stream based on Xenomai's RTnet
arvgvdevice and arvgvstream were copied and massaged to play with RTnet:
  - All relevant socket operations were moved to RTnet.
  - The heartbeat thread was omitted.
  - The data reception thread was transformed to be a function callable from
    an external application thread (arv_rt_gv_stream_recv_frame()).
  - Only UDP is supported.
There are still several NRT mechanisms which have to be eradicated in
following commits like dynamic memory allocations, obtaining Linux time,
queue and list handling including mutexes and memory allocations etc.
Nevertheless, the current state allows to create an RT GigEVision device and
to get the camera data over RTnet. The only modifications for an applications
are:
  1. Use arv_rt_gv_device_new() followed by arv_camera_new_with_device()
     instead of using arv_camera_new() (camera discovery is not supported)
  2. Call arv_rt_gv_stream_recv_frame() before arv_stream_timeout_pop_buffer().