arvrtgvstream: eradicate NRT handling of frames and buffers
Originally, every new frame was dynamically allocated and appended to a
single-linked list including further memory allocation and use of NRT
mutexes. Now a fixed set of frames is pre-allocated when the stream is
constructed. Buffers where organized in queues, also including memory
allocations and use of mutexes. The use of dynamic allocation was removed
and also no mutexes are involved any more because the asynchronous nature
of the data reception was changed to a synchronous one.

To make use of the changed behaviour towards RT an application has to
call
  - arv_rt_gv_stream_push_buffer()
  - arv_rt_gv_stream_try_pop_buffer()
  - arv_rt_gv_stream_get_n_buffers()
instead of the corresponding general arv_stream_*_buffer() functions.